+++ scripts/config/constants.config.mjs
@@ -11,4 +11,14 @@ export const SETTING_KEYS = {
   RMV_ON_COMBAT: "rmvCovCombat",
   RMV_ON_MOVE: "rmvCovMovement",
   DEBUG: "debugCover",
+  CREATURE_HEIGHTS: "creatureHeights",
+};
+
+export const DEFAULT_SIZE_FT = {
+  tiny: 1,
+  small: 3,
+  medium: 6,
+  large: 12,
+  huge: 24,
+  gargantuan: 48
 };