+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## Version 0.2.0
+
+- Support for **Sharpshooter** and **Spell Sniper** when determining whether an attack ignores cover (#1)
+- Added a dedicated **Creature Heights** settings menu
+
 ## Version 0.1.0
 
 - first release